+++ tests/test_gzip_request.js
@@ -12,8 +12,8 @@ it('should accept and decode gzip encoded application/json', done => {
 
   nock('http://example.test')
     .post('/')
-    .reply(function (url, actual) {
-      expect(actual).to.deep.equal(message)
+    .reply(async request => {
+      expect(await request.json()).to.deep.equal(message)
       done()
       return [200]
     })
@@ -29,16 +29,17 @@ it('should accept and decode gzip encoded application/json', done => {
   })
 
   const compressedMessage = zlib.gzipSync(JSON.stringify(message))
-
   req.write(compressedMessage)
   req.end()
 })
 
 it('should accept and decode gzip encoded application/json, when headers come from a client as an array', done => {
-  const compressedMessage = zlib.gzipSync(JSON.stringify({ my: 'contents' }))
+  const message = {
+    my: 'contents',
+  }
 
   const scope = nock('http://example.test')
-    .post('/', compressedMessage)
+    .post('/', message)
     .reply(200)
 
   const req = http.request({
@@ -55,6 +56,7 @@ it('should accept and decode gzip encoded application/json, when headers come fr
     done()
   })
 
+  const compressedMessage = zlib.gzipSync(JSON.stringify(message))
   req.write(compressedMessage)
   req.end()
 })
@@ -66,8 +68,8 @@ it('should accept and decode deflate encoded application/json', done => {
 
   nock('http://example.test')
     .post('/')
-    .reply(function (url, actual) {
-      expect(actual).to.deep.equal(message)
+    .reply(async request => {
+      expect(await request.json()).to.deep.equal(message)
       done()
       return [200]
     })